refactor: remove test-only prod code (audit #18 §5.2)
secretvault.New + IsLegacyEncrypted deleted (greenfield: no legacy
ciphertexts; tests use NewWithSalt); webhooks Worker.Tick wrapper deleted
(tests call tick); webhooks parseFilter moved into filter_test.go.
LatestSucceededWithContext (jobs), TrackedCount (presence),
HasOutboundSupervisor (agenttransport) relocated verbatim into _test.go —
they were test-only accessors, so moving them out of the production binary
preserves their behavioural test coverage. fleet_scope.IsAllowedAny/Filter
kept and annotated (multi-tenant stays on the roadmap).

Deviations: ConnectedAt (presence) kept in production — it has a real
cross-package test caller (server/cross_lock_presence_test.go) so it is
neither movable to presence's test binary nor deletable.
recoverSequencesFromRecords already removed by P4.
