+++ internal/controlplane/fleet/provider_redaction_test.go
@@ -24,7 +24,7 @@ func newCloudflareService(t *testing.T) *fleet.Service {
 	}
 	t.Cleanup(func() { _ = store.Close() })
 
-	vault, err := secretvault.New("test-passphrase", secretvault.AllDomains)
+	vault, err := secretvault.NewWithSalt("test-passphrase", secretvault.AllDomains, []byte("panvex-test-hkdf-salt-16b"))
 	if err != nil {
 		t.Fatalf("secretvault.New() error = %v", err)
 	}
