+++ internal/controlplane/presence/tracker.go
@@ -78,17 +78,6 @@ func (t *Tracker) Remove(agentID string) {
 	delete(t.agentTimestamps, agentID)
 }
 
-// TrackedCount returns the number of agents currently tracked (have an entry
-// in the presence map), regardless of their evaluated liveness. This is the
-// raw map size; the panvex_agent_connected gauge instead uses EvaluateAll so
-// stale-but-not-deregistered agents are excluded (L-8). Retained for
-// diagnostics/tests.
-func (t *Tracker) TrackedCount() int {
-	t.mu.RLock()
-	defer t.mu.RUnlock()
-	return len(t.agentTimestamps)
-}
-
 // ConnectedAt returns the recorded stream-open timestamp for the given agent.
 // The second return value is false when the agent is not currently tracked.
 // Used by diagnostics and tests (P2-LOG-12 / L-05) to verify that connectedAt
